Fixed CNF Generator
Turns out it was ignoring the encoding argument. Now that it takes the
encoding argument into account it turns out not all encodings support
the bounds we are looking for so many are dropped. Lastly, it is
unclear what the CNF generated for the native encoding is doing so we
will skip that also.