Splitting Out SimpleGraph Lemmas
This is not the entire namespace, but those that are independent of
the XY graph concept and will possibly end up in Mathlib. In doing
this split many global variable declarations changed in a more-or-less
arbitrary way and although it all compiles now we should spend some
time distilling which global variables belong where and which lemmas
do not depend on the current global variables.