Relocate R3y_neighbor_Ind
This inspired work to unify all proofs that there are no edges in the
vicinity of a vertex in a (3,y)-good graph but the ways this idea is
used are slightly different in each case so I could not find a
unifying theorem and ended up having two lemmas for one use-case
each. Underwhelming.

In any case, this also led to slight improvements in the statements in
SimpleGraph and some renaming.