Add Option to Solve the CNF and Make the Edge Bound Optional
Often times we are only interested in the base Ramsey encoding and not
in the edge bound so let us make it optional. This change also adds
the option to solve the resulting formula directly and output a
g6-encoded graph. For example, the witness in Inequalities.lean was
generated with

python3 e.py --solve 3 7 19

The dependency on networkx is kind of extreme and should probably be
folded into the part of the code that really needs it. Furthermore,
there should be a simple way to support solving the PB formula as
well.